Refactored the Read/Write process to simplify the implementation of new file formats.
.parse/.format method are implemented at configuration level (not configuration value).

+++ test/simple.py
@@ -39,4 +39,5 @@ class TestSimpleConfig(object):
 			assert(len(list(cfg.read_values())) == 5)
 
 		finally:
-			os.remove(t)
+			if os.path.exists(t):
+				os.remove(t)
